chg: Upate container base image and publish latest version on DockerHub
Update the base image to use latest Ubuntu base image and use the /udpgen binary as entrypoint. When no argument is set
the udpgen tool gives a a help as output. To avoid mistakes we just print the help on container start and don't generate
UDP traffic accidentally.

Fixed the readme typo for -S to indicate a given limit by seconds